Version finale

+++ src/Controller/ImprimanteController.php
@@ -17,6 +17,9 @@ class ImprimanteController extends AbstractController
     #[Route('/imprimantes', name: 'app_imprimante')]
     public function index(Request $request, ImprimanteRepository $imprimanteRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $imprimantes = $imprimanteRepository->findAll();
 
         return $this->render('imprimante/index.html.twig', [
@@ -27,6 +30,9 @@ class ImprimanteController extends AbstractController
     #[Route('/imprimantes/ajout', name: 'app_imprimante_ajout')]
     public function ajouter(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $imprimante = new Imprimante();
         $form = $this->createForm(ImprimanteType::class, $imprimante);
         $form->handleRequest($request);
@@ -47,6 +53,9 @@ class ImprimanteController extends AbstractController
     #[Route('/imprimantes/{id}', name: 'app_imprimante_details', requirements: ['id' => '\d+'])]
     public function details(Imprimante $imprimante): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         return $this->render('imprimante/details.html.twig', [
             'imprimante' => $imprimante,
         ]);
@@ -55,6 +64,9 @@ class ImprimanteController extends AbstractController
     #[Route('/imprimante/{id}/supprimer', name: 'app_imprimante_supprimer')]
     public function supprimer(Imprimante $imprimante, EntityManagerInterface $entityManager): RedirectResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         if ($this->getUser()->getType() === 'Administration' || $this->getUser()->getNiveau() === 3) {
             $entityManager->remove($imprimante);
             $entityManager->flush();
